Break out api endpoint handlers
